refactor(mypage): Translate useMypage
